Move Store interfaces to Store namespace
