+++ src/Microsoft.AspNet.Identity.EntityFramework/RoleStore.cs
@@ -8,6 +8,7 @@ using System.Linq;
 using System.Security.Claims;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.AspNet.Identity.Store;
 using Microsoft.Data.Entity;
 using Microsoft.Data.Entity.Update;
