+++ src/Microsoft.AspNet.Identity/IdentityBuilder.cs
@@ -2,6 +2,7 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
+using Microsoft.AspNet.Identity.Store;
 using Microsoft.Framework.DependencyInjection;
 
 namespace Microsoft.AspNet.Identity
