+++ src/Microsoft.AspNet.Identity/Store/IUserLockoutStore.cs
@@ -5,7 +5,7 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Microsoft.AspNet.Identity
+namespace Microsoft.AspNet.Identity.Store
 {
     /// <summary>
     /// Provides an abstraction for a storing information which can be used to implement account lockout,
