+++ src/Microsoft.AspNet.Identity/Store/IUserLoginStore.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Microsoft.AspNet.Identity
+namespace Microsoft.AspNet.Identity.Store
 {
     /// <summary>
     /// Provides an abstraction for storing information that maps external login information provided
