+++ src/Microsoft.AspNet.Identity/Store/IUserPhoneNumberStore.cs
@@ -4,7 +4,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Microsoft.AspNet.Identity
+namespace Microsoft.AspNet.Identity.Store
 {
     /// <summary>
     /// Provides an abstraction for a store containing users' telephone numbers.
