+++ src/Microsoft.AspNet.Identity/Store/IUserTwoFactorStore.cs
@@ -4,7 +4,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Microsoft.AspNet.Identity
+namespace Microsoft.AspNet.Identity.Store
 {
     /// <summary>
     /// Provides an abstraction to store a flag indicating whether a user has two factor authentication enabled.
