+++ src/Microsoft.AspNet.Identity/UserManager.cs
@@ -13,6 +13,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNet.Hosting;
 using Microsoft.AspNet.Http;
 using Microsoft.AspNet.Identity.Logging;
+using Microsoft.AspNet.Identity.Store;
 using Microsoft.Framework.Logging;
 using Microsoft.Framework.OptionsModel;
