+++ test/Microsoft.AspNet.Identity.EntityFramework.Test/UserStoreGuidKeyTest.cs
@@ -2,6 +2,7 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
+using Microsoft.AspNet.Identity.Store;
 using Microsoft.Framework.DependencyInjection;
 using Xunit;
