+++ test/Microsoft.AspNet.Identity.InMemory.Test/HttpSignInTest.cs
@@ -8,6 +8,7 @@ using Microsoft.AspNet.Builder.Internal;
 using Microsoft.AspNet.Hosting;
 using Microsoft.AspNet.Http;
 using Microsoft.AspNet.Http.Authentication;
+using Microsoft.AspNet.Identity.Store;
 using Microsoft.AspNet.Identity.Test;
 using Microsoft.Framework.DependencyInjection;
 using Microsoft.Framework.Runtime.Infrastructure;
