+++ test/Microsoft.AspNet.Identity.InMemory.Test/InMemoryStoreTest.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Linq.Expressions;
+using Microsoft.AspNet.Identity.Store;
 using Microsoft.AspNet.Identity.Test;
 using Microsoft.Framework.DependencyInjection;
