+++ test/Microsoft.AspNet.Identity.Test/IdentityBuilderTest.cs
@@ -6,6 +6,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.AspNet.Identity.Store;
 using Microsoft.Framework.DependencyInjection;
 using Xunit;
