+++ test/Microsoft.AspNet.Identity.Test/NoopUserStore.cs
@@ -1,9 +1,9 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.AspNet.Identity.Store;
 
 namespace Microsoft.AspNet.Identity.Test
 {
